backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
torch
torchvision
pillow==10.1.0
python-multipart==0.0.6
